scripts/prerequisites/requirements.txt
typer~=0.12.5
rich~=13.8.1
kubernetes~=30.1.0
toml~=0.10.2
requests~=2.32.3
xmltodict~=0.13.0
PyYAML~=6.0.2
ruamel.yaml==0.18.6
tomlkit~=0.13.2
ldap3~=2.9.1
cryptography~=43.0.1
docker~=7.1.0
pyOpenSSL~=24.2.1
podman==5.0.0